* Moved Document type files into sub-folder to mirror layout of functions and domainlogic packages.
